Add schematizable EDN validator (#6)
[Card](https://rentpath.atlassian.net/browse/SRV-1083)

Implements an extended EDN Serde (based on the native jackdaw Serde), with an additional option to validate messages against the required rules for Kafka schema autoderivation (specifically in the context of Kafka Connect). By setting the `:validate?` option to true, an error will be thrown before an invalid message is published to kafka, and structs anywhere within the message will be automatically homogenized.