Add default key and value serdes (#12)
This adds precompiled classes for reasonable default configurations for
key and value serdes. This is necessary as the KStream API requires that
default serdes be specified as strings representing class names on the
class path, and cannot be automatically generated from a function. The
same does not apply to per-topic serdes.