Ensure space preceding set literal
[Card](https://rentpath.atlassian.net/browse/SRV-1117)

Because the Clojure EDN reader interprets the hash mark in a string
such as the following: `{:somekey#{1 2 3}}` as being a part of the
preceding symbol (i.e. `:somekey#`), rather than an indicator that the
following data structure is a set (and not a map), it then tries to read
the subsequent structure as a map and fails (as it has an odd number of
values). This ensures that in the compact EDN format every set will be
serialized with a preceding space if one is not already enforced
otherwise.